Update method to reset projection
For now, clearing the projection region and reading the source data from the input file remains the simplest and most performant implementation for resetting the projection region for projection. If we need a more robust approach we could copy the node-set and mesh directly from the source region to the projection region.